clean tree folder
