+++ format/src/main/java/com/github/sadikovi/riff/tree/node/IsNull.java
@@ -20,38 +20,25 @@
  * SOFTWARE.
  */
 
-package com.github.sadikovi.riff.tree.expression;
+package com.github.sadikovi.riff.tree.node;
 
 import org.apache.spark.sql.catalyst.InternalRow;
 
 import com.github.sadikovi.riff.ColumnFilter;
 import com.github.sadikovi.riff.Statistics;
+import com.github.sadikovi.riff.tree.BoundReference;
 import com.github.sadikovi.riff.tree.Rule;
 import com.github.sadikovi.riff.tree.Tree;
-import com.github.sadikovi.riff.tree.TypedBoundReference;
-import com.github.sadikovi.riff.tree.TypedExpression;
 
 /**
- * [[GreaterThanOrEqual]] is inequality predicate for typed expression.
- * Ordinal row value is greater than or equal to expression value.
+ * [[IsNull]] node represents field that has evaluated against null values. When row has a null
+ * value for this node's field, yields true, otherwise false. Does not have information about type.
  */
-public class GreaterThanOrEqual extends TypedBoundReference {
+public class IsNull extends BoundReference {
   private final String name;
-  private final TypedExpression expr;
 
-  public GreaterThanOrEqual(String name, TypedExpression expr) {
+  public IsNull(String name) {
     this.name = name;
-    this.expr = expr;
-  }
-
-  @Override
-  public TypedExpression expression() {
-    return this.expr;
-  }
-
-  @Override
-  public String operator() {
-    return ">=";
   }
 
   @Override
@@ -61,17 +48,16 @@ public class GreaterThanOrEqual extends TypedBoundReference {
 
   @Override
   public boolean evaluateState(InternalRow row, int ordinal) {
-    return !row.isNullAt(ordinal) && expr.geExpr(row, ordinal);
+    return row.isNullAt(ordinal);
   }
 
   @Override
   public boolean evaluateState(Statistics stats) {
-    return !stats.isNullAt(Statistics.ORD_MAX) && expr.geExpr(stats, Statistics.ORD_MAX);
+    return stats.hasNulls();
   }
 
   @Override
   public boolean evaluateState(ColumnFilter filter) {
-    // column filter is not evaluated for GreaterThanOrEqual
     return true;
   }
 
@@ -82,6 +68,11 @@ public class GreaterThanOrEqual extends TypedBoundReference {
 
   @Override
   public Tree copy() {
-    return new GreaterThanOrEqual(name, expr.copy()).copyOrdinal(this);
+    return new IsNull(name).copyOrdinal(this);
+  }
+
+  @Override
+  public String toString() {
+    return prettyName() + " is null";
   }
 }
