+++ format/src/main/java/com/github/sadikovi/riff/tree/node/Not.java
@@ -20,59 +20,57 @@
  * SOFTWARE.
  */
 
-package com.github.sadikovi.riff.tree.expression;
+package com.github.sadikovi.riff.tree.node;
 
 import org.apache.spark.sql.catalyst.InternalRow;
 
 import com.github.sadikovi.riff.ColumnFilter;
 import com.github.sadikovi.riff.Statistics;
 import com.github.sadikovi.riff.tree.Rule;
+import com.github.sadikovi.riff.tree.State;
 import com.github.sadikovi.riff.tree.Tree;
-import com.github.sadikovi.riff.tree.TypedBoundReference;
-import com.github.sadikovi.riff.tree.TypedExpression;
+import com.github.sadikovi.riff.tree.UnaryLogical;
 
 /**
- * [[LessThanOrEqual]] is inequality predicate for typed expression.
- * Ordinal row value is less than or equal to expression value.
+ * [[Not]] is unary logical node that inverses value of the child tree. Note that in some cases
+ * we keep `true` instead of inversing child's value - this is done to avoid false reporting of
+ * filters.
  */
-public class LessThanOrEqual extends TypedBoundReference {
-  private final String name;
-  private final TypedExpression expr;
+public class Not extends UnaryLogical {
+  private final Tree child;
 
-  public LessThanOrEqual(String name, TypedExpression expr) {
-    this.name = name;
-    this.expr = expr;
+  public Not(Tree child) {
+    this.child = child;
   }
 
   @Override
-  public TypedExpression expression() {
-    return this.expr;
+  public Tree child() {
+    return child;
   }
 
   @Override
-  public String operator() {
-    return "<=";
+  public boolean evaluateState(InternalRow row) {
+    return !child.evaluateState(row);
   }
 
   @Override
-  public String name() {
-    return name;
-  }
-
-  @Override
-  public boolean evaluateState(InternalRow row, int ordinal) {
-    return !row.isNullAt(ordinal) && expr.leExpr(row, ordinal);
-  }
-
-  @Override
-  public boolean evaluateState(Statistics stats) {
-    return !stats.isNullAt(Statistics.ORD_MIN) && expr.leExpr(stats, Statistics.ORD_MIN);
-  }
-
-  @Override
-  public boolean evaluateState(ColumnFilter filter) {
-    // column filter is not evaluated for LessThanOrEqual
+  public boolean evaluateState(Statistics[] stats) {
+    // `not` does not evaluate statistics, because child expression evaluation does not guarantee
+    // that records for inverse result will not exist in dataset
     return true;
+  }
+
+  @Override
+  public boolean evaluateState(ColumnFilter[] filters) {
+    // `not` does not evaluate column filters, because child expression evaluation does not
+    // guarantee that records for inverse result will not exist in dataset
+    return true;
+  }
+
+  @Override
+  public State state() {
+    // state of not predicate is unknown similar to evaluation
+    return State.Unknown;
   }
 
   @Override
@@ -82,6 +80,11 @@ public class LessThanOrEqual extends TypedBoundReference {
 
   @Override
   public Tree copy() {
-    return new LessThanOrEqual(name, expr.copy()).copyOrdinal(this);
+    return new Not(child.copy());
+  }
+
+  @Override
+  public String toString() {
+    return "!(" + child + ")";
   }
 }
